requirements.txt
aiogram
aiofiles
openai
python-dotenv
